Add createExpoTargetConfig
