+++ app.plugin.js
@@ -23,7 +23,6 @@ const withActivityMonitorExtensionPlugin = (config, props) => {
       ),
       props,
     ),
-
     props,
   );
 };
